// src/model/Poke.ts
export default interface Poke {
    abilities:                Ability[];
    base_experience:          number;
    forms:                    Species[];
    game_indices:             GameIndex[];
    height:                   number;
    held_items:               HeldItem[];
    id:                       number;
    is_default:               boolean;
    location_area_encounters: string;
    moves:                    Move[];
    name:                     string;
    order:                    number;
    species:                  Species;
    sprites:                  Sprites;
    stats:                    Stat[];
    types:                    Type[];
    weight:                   number;
}

interface Ability {
    ability:   Species;
    is_hidden: boolean;
    slot:      number;
}

interface Species {
    name: string;
    url:  string;
}

interface GameIndex {
    game_index: number;
    version:    Species;
}

interface HeldItem {
    item:            Species;
    version_details: VersionDetail[];
}

interface VersionDetail {
    rarity:  number;
    version: Species;
}

interface Move {
    move:                  Species;
    version_group_details: VersionGroupDetail[];
}

interface VersionGroupDetail {
    level_learned_at:  number;
    move_learn_method: Species;
    version_group:     Species;
}

interface GenerationV {
    "black-white": Sprites;
}

interface GenerationIv {
    "diamond-pearl":        Sprites;
    "heartgold-soulsilver": Sprites;
    platinum:               Sprites;
}

interface Versions {
    "generation-i":    GenerationI;
    "generation-ii":   GenerationIi;
    "generation-iii":  GenerationIii;
    "generation-iv":   GenerationIv;
    "generation-v":    GenerationV;
    "generation-vi":   { [key: string]: GenerationVi };
    "generation-vii":  GenerationVii;
    "generation-viii": GenerationViii;
}

interface Sprites {
    back_default:       string;
    back_female:        string;
    back_shiny:         string;
    back_shiny_female:  string;
    front_default:      string;
    front_female:       string;
    front_shiny:        string;
    front_shiny_female: string;
    other?:             Other;
    versions?:          Versions;
    animated?:          Sprites;
}

interface GenerationI {
    "red-blue": RedBlue;
    yellow:     RedBlue;
}

interface RedBlue {
    back_default:  string;
    back_gray:     string;
    front_default: string;
    front_gray:    string;
}

 interface GenerationIi {
    crystal: Crystal;
    gold:    Crystal;
    silver:  Crystal;
}

 interface Crystal {
    back_default:  string;
    back_shiny:    string;
    front_default: string;
    front_shiny:   string;
}

 interface GenerationIii {
    emerald:             Emerald;
    "firered-leafgreen": Crystal;
    "ruby-sapphire":     Crystal;
}

 interface Emerald {
    front_default: string;
    front_shiny:   string;
}

 interface GenerationVi {
    front_default:      string;
    front_female:       string;
    front_shiny:        string;
    front_shiny_female: string;
}

 interface GenerationVii {
    icons:                  DreamWorld;
    "ultra-sun-ultra-moon": GenerationVi;
}

 interface DreamWorld {
    front_default: string;
    front_female:  string;
}

 interface GenerationViii {
    icons: DreamWorld;
}

 interface Other {
    dream_world:        DreamWorld;
    "official-artwork": OfficialArtwork;
}

 interface OfficialArtwork {
    front_default: string;
}

 interface Stat {
    base_stat: number;
    effort:    number;
    stat:      Species;
}

 interface Type {
    slot: number;
    type: Species;
}